+++ doc/examples/particles/lib/main.dart
@@ -63,6 +63,12 @@ class MyGame extends BaseGame {
     Timer.periodic(sceneDuration, (_) => spawnParticles());
   }
 
+  @override
+  Future<void> onLoad() async {
+    await images.load('zap.png');
+    await images.load('boom3.png');
+  }
+
   /// Showcases various different uses of [Particle]
   /// and its derivatives
   void spawnParticles() {
@@ -293,7 +299,7 @@ class MyGame extends BaseGame {
   Particle imageParticle() {
     return ImageParticle(
       size: Vector2.all(24),
-      image: Flame.images.loadedFiles['zap.png'].loadedImage,
+      image: images.fromCache('zap.png'),
     );
   }
 
@@ -383,7 +389,7 @@ class MyGame extends BaseGame {
   /// Flame's [Sprite] into the effect.
   Particle spriteParticle() {
     return SpriteParticle(
-      sprite: Sprite('zap.png'),
+      sprite: Sprite(images.fromCache('zap.png')),
       size: cellSize * .5,
     );
   }
@@ -528,12 +534,11 @@ class MyGame extends BaseGame {
     const columns = 8;
     const rows = 8;
     const frames = columns * rows;
-    const imagePath = 'boom3.png';
-    final spriteImage = Flame.images.loadedFiles[imagePath].loadedImage;
+    final spriteImage = images.fromCache('boom3.png');
     final spritesheet = SpriteSheet(
       rows: rows,
       columns: columns,
-      imageName: imagePath,
+      image: spriteImage,
       textureWidth: spriteImage.width ~/ columns,
       textureHeight: spriteImage.height ~/ rows,
     );
@@ -547,19 +552,9 @@ class MyGame extends BaseGame {
 }
 
 Future<BaseGame> loadGame() async {
-  Vector2 gameSize;
+  Flame.initializeWidget();
+  final gameSize = await Flame.util.initialDimensions();
   WidgetsFlutterBinding.ensureInitialized();
-
-  await Future.wait([
-    Flame.util.initialDimensions().then((size) => gameSize = size),
-    Flame.images.loadAll(const [
-      'zap.png',
-
-      /// Credits to Stumpy Strust from
-      /// https://opengameart.org/content/explosion-sheet
-      'boom3.png',
-    ]),
-  ]);
 
   return MyGame(screenSize: gameSize);
 }