+++ doc/examples/spritesheet/lib/main.dart
@@ -2,57 +2,26 @@ import 'package:flame/extensions/vector2.dart';
 import 'package:flutter/material.dart';
 import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flame/components/sprite_component.dart';
-import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flame/spritesheet.dart';
-import 'package:dashbook/dashbook.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-  final spriteSheet = SpriteSheet(
-    imageName: 'spritesheet.png',
-    textureWidth: 16,
-    textureHeight: 18,
-    columns: 11,
-    rows: 2,
-  );
-
-  final spriteSheetFromImage = SpriteSheet.fromImage(
-    image: await Flame.images.load('spritesheet.png'),
-    textureWidth: 16,
-    textureHeight: 18,
-    columns: 11,
-    rows: 2,
-  );
-
-  final dashbook = Dashbook();
-
-  dashbook
-      .storiesOf('SpriteSheet')
-      .add('defaut', (_) => GameWrapper(MyGame(spriteSheet)))
-      .add('fromImage', (_) => GameWrapper(MyGame(spriteSheetFromImage)));
-
-  runApp(dashbook);
-}
-
-class GameWrapper extends StatelessWidget {
-  final Game game;
-
-  GameWrapper(this.game);
-
-  @override
-  Widget build(_) {
-    return Container(
-      width: 400,
-      height: 400,
-      child: game.widget,
-    );
-  }
+  runApp(MyGame().widget);
 }
 
 class MyGame extends BaseGame {
-  MyGame(SpriteSheet spriteSheet) {
+  @override
+  Future<void> onLoad() async {
+    final spriteSheet = SpriteSheet(
+      image: await images.load('spritesheet.png'),
+      textureWidth: 16,
+      textureHeight: 18,
+      columns: 11,
+      rows: 2,
+    );
+
     final vampireAnimation =
         spriteSheet.createAnimation(0, stepTime: 0.1, to: 7);
     final ghostAnimation = spriteSheet.createAnimation(1, stepTime: 0.1, to: 7);